refactor: merge 'tests' and 'test-engineering' directories
