+++ tests/contract/client/tests/exceptions.py
