+++ tests/contract/client/tests/kinto.py
