+++ tests/contract/kinto-attachments/.gitignore
@@ -1,2 +1,2 @@
-*
+main
 !.gitignore
