+++ tests/contract/kinto-setup/Dockerfile
