+++ tests/contract/kinto-setup/README.md
