+++ tests/contract/kinto-setup/dev-requirements.in
