+++ tests/contract/kinto-setup/kinto.py
