+++ tests/contract/kinto-setup/main.py
