+++ tests/contract/kinto-setup/mypy.ini
