+++ tests/contract/kinto-setup/requirements.in
