+++ tests/contract/kinto-setup/requirements.txt
