+++ tests/contract/volumes/kinto/data-03.json
