+++ tests/contract/volumes/kinto/data-04.json
