+++ tests/load/setup_k8s.sh
