+++ tests/unit/web/test_api_v1.py
@@ -3,8 +3,8 @@ from fastapi.testclient import TestClient
 
 from merino.main import app
 from merino.providers import get_providers
-from tests.web.util import filter_caplog
-from tests.web.util import get_providers as override_dependency
+from tests.unit.web.util import filter_caplog
+from tests.unit.web.util import get_providers as override_dependency
 
 client = TestClient(app)
